backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# LLM — Groq (free tier: llama-3.3-70b-versatile)
groq>=0.12.0

# File parsers
python-docx>=1.1.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
pypdf>=3.17.0
pandas>=2.1.0
Pillow>=10.1.0

# OCR for image evidence (requires Tesseract binary — see README)
pytesseract>=0.3.10

# CLI
click>=8.1.0

# Templating (HTML reports)
jinja2>=3.1.0

# Multipart (file uploads)
python-multipart>=0.0.7
